when a build is attempted, don't save version
it's weird to have a build's inputs half-filled-in. the downside of this
change is that when N versions are detected, it will take as long as
(N*fetches) to trigger them all, rather than immediately triggering as the
versions can be seen to be different.

the upside is simplicity in the interim until this is all redone.