builds api speaks generic build plans
this is a backwards-compatible change that paves the way for jobs having
generic build plans composing many actions arbitrarily.

the current 'inputs -> config -> outputs' config style will always be
supported, as that's the very common case (and is easy to translate).

note: we elected to not version the exec engine for this change (we're still
pre-1.0), which only means that builds running *during* the deploy will
probably error out.

[#88383920]

Signed-off-by: Chris Brown <cbrown@pivotal.io>