initial rework of UI to support build plans
all this does is process old events in a new way, that will be easy to add
arbitrary build plan support for in the future.

this sacrifices the pre-drawing of build inputs, which isn't really a thing
anymore with arbitrary build plans. something to consider adding back in the
future.