'put' steps receive full repository as input
this (backwards-INcompatible) change makes the 'inputs' to a 'put' step
more explicit. rather than willy-nilly taking the bits from the previous
step, every source in the repository is propagated to the 'put' step,
under its given name.

this means that things that previously looked like this:

    plan:
    - get: some-repo
    - task: commit-to-repo
    - put: some-other-repo
      params:
        repository: some-repo

...now looks like this:

    plan:
    - get: some-repo
    - task: commit-to-repo
    - put: some-other-repo
      params:
        repository: commit-to-repo/some-repo

in the future, rather than fetching *all* sources from the repository,
there may be a way to explicitly satisfy a 'put' action's input
dependencies.

this removes the need for all Steps to be able to stream to the next
step; this will be cleaned up in a later commit.

[#90339654]