refactor volume management for better reattaching
previously, when reattaching to a build, and resuming the steps, we
would pick a worker, see if the resource exists on that worker, if so
attach, and if not create. this regressed from the previous behavior of,
find a container *globally*, and if that doesn't exist create.

this pushes most of the cache handling logic down into `resource/` and
greatly cleans up the Get step at the same time.

[#104812656]