basic auth and github auth can both be specified
If both are configured:

* API endpoints permit both and reject w/ basic flow (rather than /login
  redirect)
* Web endpoints permit both and reject w/ redirect to /login

If either are exclusively configured, the auth is used for both.

Along the way cleaned up the session handler to be more durable to bogus
sessions flowing through (i.e. if the secret changes due to changing
config).

Added some validation to the CLI to verify that -externalURL is
specified if github auth client info is specified.

[#105045438]