do not implicitly create touched jobs
[#110821444]

Another boring commit.

This change removes the concept of lazily creating resources at the
point when they are first accessed. We now create them up front when the
pipeline configuration that contains them is saved. This will reduce the
amortised cost of checking if a job exists at unexpected points and
reduce the roundtrips to the database.

* Update all of the database integration tests so that before they use a
  job they save a pipeline configuration with that job first. This
  mainly involved standardising on using `some-job` and `some-other-job`
  for resource names in our tests.

* The `GetJobs()` function has gone. It was only used to test that
  resources were saved when their pipeline configuration was saved. Now
  that `GetJob()` has no side effects we can use that instead.