Reduce CPU usage for animations
Update CSS to use transforms in favor of box-shadow manipulations to
create ripple animations. By using only CSS transforms the amount of
repainting required by the browser is reduced. Browsers can now
render the animations on a different layer and make use of the
GPU [1].

The legend has also been updated to remove all animation. Whilst not
as pretty, this also helps to reduce CPU consumption.

[1] http://www.paulirish.com/2012/why-moving-elements-with-translate-is-better-than-posabs-topleft/