return false if volume is not in the db
previously this would return ErrMissingVolume, which was a total lie.
that error is for when a volume which we absolutely expect to exist, as
it's mounted to a container, is no longer there. this led to funky
control flow and ultimately funky logging.

this also resulted in removing an 'if !found || err != nil', which is
just about always a bad idea