mark job/resource as active when updating config
Instead of pulling pipeline config every time on read requests, just get
active resources, resource types and jobs from database. When pipeline
config is updated it will mark all jobs, resources and resource types
that are no longer in config as inactive.

We know we updated existing migration but we did not push yet the
migration so we think it is fine.

[#127510245]

Signed-off-by: Maria Shaldibina <mshaldibina@pivotal.io>