fix another 100%-cpu case in the algorithm
the actual fix: 'break' -> 'return nil, false'

I added a bunch of hooks for adding metrics, which helped out with
debugging this. They'll only be sent if env vars are configured for
talking to InfluxDB.

Here's what they look like, before and after the fix:

    http://i.imgur.com/2MEuOUM.png

I've also checked in the test and data that repro'd it, and checked in
the previous data + test that repro'd the previous CPU issue. Test data
is now GZip-compressed, which makes this viable.