when trying to look up container also check worker is running
FindContainerByIdentifier did not have to deal with worker states
because it existed before worker state was introduced. We use
FindContainerByIdentifier to lookup existing container and re-attach to it
regardless of what worker it is in and then use the worker it is in. This is
opposite to when we pick worker first and then finding and creating container
on it in any other case.

[#140167681]