reduce usage of container identifier via 'user'
the *feature* that this accomplishes is to remove check_source and
check_type from the database. the rest is a bunch of refactoring in the
interest of removing container identifier in the long run.

we wanted to just kill things from container identifier, but then saw
that it was used to correlate containers (for e.g. fetching
image_resource, image_resource_type) with each other by building on the
same ID, and using the ID to determine if the container is for a build,
resource, or resource type.

we introduced a dbng.ResourceUser interface which removes this usage
pattern of the identifier, making it easier to remove later.
dbng.ResourceUser is provided alongside the identifier from the very top
of the call site (where it knows that it's for a build, or resource,
etc.). it is then passed along all the way down the stack, and
everything that us 'used' in the database is done via the ResourceUser.
this also allowed us to remove many 'XForBuild', 'XForResource',
'XForResourceType' methods and objects, and consolidate them into one
that takes a dbng.ResourceUser.

we should be able to remove the container identifier entirely, once we
figure out all the little subtle things it was being used for.

[#139984521]

Signed-off-by: Clara Fu <cfu@pivotal.io>