always create container if there is no initialized volume
if atc gets interrupted in the middle of initialization there would be
uninitialized volume and container for it. We should start fresh in that
case and create new volume and container for it letting gc taking care
of duplicates.

VolumeCollector#GetDuplicateResourceCacheVolumes is already removing
uninitialized volumes if there is initialized one for the same resource
cache on the same worker. Containers will be cleaned up when build is
finished

[#140566469]