remove 'Ephemeral'
back in the old days this was used to nuke ephemeral containers on
worker shutdown, via the jettison component. this was originally to
force recreating of the containers so new versions of the base resource
types would be used. this is long gone and life explicitly manages
versions, so this is no longer needed.