default sslmode to disable
we defaulted to this before (by way of the datasource flag having
?sslmode=disable), but changed it to 'verify-full' in the interest of
having secure defaults. however, this is hella backwards-incompatible,
and immediately broke CI, and will break the Vagrant images as well. we
can't really know how the binaries will be deployed, and if SSL will
even be necessary, if e.g. postgres is just listening on a local socket.

what we really want here is 'prefer', but lib/pq doesn't support it.

Signed-off-by: Alex Suraci <asuraci@pivotal.io>