rip out pgx in favor of regular ol pq
we originally used pgx so we could guarantee only one connection was
used for locking, as locking is connection-based. this introduced a
whole slew of different problems, though; pgx does not support
configuring certs, has nonstandard parsing of the datasource connection
string, and users have run into incompatibilities between it and pq in
the past.

so, instead, use a *sql.DB limited to one connection. this should also
handle retrying, so we can kill *RetryableConn.

Signed-off-by: Clara Fu <cfu@pivotal.io>