fix create resource config/cache for deleted user
previously, any foreign_key_violation error on insertig the use would be
retried. this would result in an infinite retry loop in the case that
the foreign_key_violation is actually for the user (i.e. build,
resource, etc) - for example, if a pipeline or team is destroyed while
the find-or-create is happening.

this infinite loop would also tend to happen while a lock was held,
resulting in deadlocked builds.

the foreign_key_violation is now checked to be regarding the specific
column we care about, and tests have been backfilled to cover the
original intent (concurrent removal of the cache/config being used), in
addition to a test for the user being removed.

[#144164973]