add compatibility checking for workers
Before a new garden connection is made (or reattached) to a worker,
there will be an actual worker version vs wanted worker version
compatibility check. The worker version is only compatible if the
major version is the same and if the minor version is either equal
or greater than the wanted minor version.

concourse/concourse#1047

Signed-off-by: Clara Fu <cfu@pivotal.io>