fix hanging postgres on db/acceptance suite exit
a bump to ubuntu seems to have made it so 'su' no longer replaces the
current process with the command it's running, leading to the SIGINT
killing the 'su' and leaving the postgres owned by init.

or, alternatively, it's always done this and has stopped propagating
'SIGINT' to its children or something.

so instead, just use Go APIs for setting the user, which is much cleaner
anyway

Signed-off-by: Andrew Edstrom <aedstrom@pivotal.io>