show metadata when resource cache is used
concourse/concourse#932

* add migration to add metadata to resource cache + migrate all the data
We keeping metadata in versioned_resources still cause resource caches
only exist for latest resource versions (it is displayed on resource
page). Essentially, we want metadata to be stored on versioned_resources.
Currently, we get it after we fetch resource. In case if we find
resource cache fetch does not happen, so we store it in resource cache
after fetch and update versioned_resources with metadata. There is a proposal to get
it when resource checks are run
(https://github.com/concourse/concourse/issues/534)
In that case we won't need to store it in resource cache.

* refactor resource.Fetcher to return VersionedSource and not
FetchSource since we only need VersionedSource. This allows us to get
the versionedSource as a result of the function instead of having it on
a FetchSource as partially instantiated.

* remove pipelineDB methods that set metadata and only leave
dbng/pipeline to do that.

* remove unused ContainerFetchSource

Signed-off-by: Andrew Edstrom <aedstrom@pivotal.io>