obtain resource cache metadata under the lock
If two processes fetch the same resource the resource cache metadata
will be empty before we check if it is initialized, etc. So one of the
processes might have an outdated resourceCache with empty metadata which
will be filled in by another process. So instead get metadata in
FindInitialized which runs under the lock and return the metadata that
was populated by another process.

concourse/concourse#932

Signed-off-by: Maria Shaldibina <mshaldibina@pivotal.io>