creates initial path for finding or creating a container
A container owner is introduced in order to create a path for
FindOrCreateContainer instead of multiple find or create methods
that are specific to a build, get, or check. This general method
will take the container owner in order to determine the columns
to set or query for.

concourse/concourse#1242

Signed-off-by: Alex Suraci <asuraci@pivotal.io>