Filter resource versions by type.
If the type of a resource changes, resource versions from the previous
resource type can be fetched and passed to builds. This leads to
unexpected build errors, since resource version metadata from one
resource type may not be usable to another resource type. This commit
filters resource versions by the current type of the resource.