Don't cache badges; they might get stale
Per beyhan:

> Github has aggressive image caching which breaks CI badges. See [issue](https://github.com/github/markup/issues/224).
> Concourse doesn't set the Cache-Control header to no-cache which is
> expected by Github.

We set the `Cache-Control` headers to `no-cache, no-store, must-revalidate`
because that's what it's set to in [this commit](https://github.com/concourse/atc/commit/3740efa),
and we want to be consistent because, after all, "consistency is the
hobgoblin of small minds".

A longer explanation of the various cache controls can be found
[here](https://stackoverflow.com/questions/49547/how-to-control-web-page-caching-across-all-browsers).

We've chosen not to implement all the controls because they are not
relevant to the audience who is viewing Concourse badges on GitHub (e.g.
they are not using Internet Explorer 6), and, as aforementioned, we'd
like to be consistent with the rest of the codebase.

[fixes concourse/concourse#835]
[finishes #136613913]