fix ever-growing chain of causality from circles
our pipeline has an 'rc' job, which takes the current 'version' and
outputs a bumped 'version'. following the outputs, we will eventually
arrive at the next build of 'rc', and everything *that* leads to. this
meant that viewing commit N-3 of concourse would include the causality,
and then N-2's causality, and the N-1, and then N.

as a workaround for this case, we can restrict the resulting set to only
include the earliest version of each resource found.

there may be some cases where this is imperfect, for example if a
pipeline does a bunch of 'put's to the same resource throughout a normal
run. a strange example may be notifications, but there may be more
practical examples as well. we'll have to see, and possibly just remodel
things so this query isn't so complicated.

concourse/concourse#1538