a better approach to improving cycles
the previous approach had a few downsides:

* in cases where the same resource was updated repeatedly through a
  single run, only the first version would be reported
* in cases where the cycle led to more jobs running in one iteration of
  the cycle (e.g. in our pipeline, shipping a candidate, leading to
  various publishing jobs running), the result would still include
  those jobs

the new approach is to only select inputs that are not also outputs of a
previous build of the same job. this will fix cycles introduced by jobs
like 'rc' that fetch a version X.X.X-rc.N and push X.X.X-rc.(N+1) (which
then feeds in to the next 'rc').

concourse/concourse#1538