optimize dashboard via materlialized views
this introduces a performance trade-off: rather than computing these
queries every time the dashboard refreshes (every 5 seconds), we
compute them whenever a build starts or completes (depending on the
view).

this means that performance may be *worse* if builds are starting and
finishing more often.

in a later iteration we'll make the refreshes use `CONCURRENTLY`,
however this feature does not exist in Postgres 9.3, which our BOSH
release is stuck with. we'll be switching off in the next release.

a even later iteration may be to batch the refreshing to once per N
seconds (1?), and change the existing `REFRESH` calls to instead set a
flag for this fixed interval to check whether a refresh is necessary.

concourse/concourse#1589
concourse/concourse#1555

Signed-off-by: Gabriel Dumitrescu <gdumitrescu@pivotal.io>