Cache ACME certificates in database
LetsEncrypt rate limits duplicate certificates to 5 per week.
If atc were to be restarted 6 times in a 1 week period, it would fail to start
the 6th time due to not being able to obtain a certificate from LetsEncrypt.
This commit includes a db migration to add a table in the database `cert_cache`
which is used to save certificates.
It then uses this table as a cache to avoid hitting LetsEncrypt on startup.