fix piling up of database connections due to lock
previously, we would open a transaction, bump the scheduling time, and
then with the transaction still held open, acquire the scheduling lock.

acquiring the lock is intended to not block, but we observed it hanging
in production. this isn't too surprising; fundamentally we should never
open a transaction and leave it open while depending on other I/O that's
not within the same transaction.

instead, we'll now immediately try to acquire the lock, and *then* bump
the timestamp, which can now happen on its own connection. if the
timestamp doesn't bump, the lock is released.

concourse/concourse#1734