exec, resource: switch to context.Context
this makes the interface a bit more straightforward. context.Context has
cancellation and timeout semantics that are pretty much a perfect fit
for aborting and the 'timeout' step, and allows the tests to easily use
a synchronous interface.

along the way, remove exec.ErrInterrupted and resource.ErrAborted, which
are replaced by context.Canceled.