Create credhub client lazily.
Creating a credhub client can fail if credhub isn't available
(specifically if we pass options that require connecting to credhub). To
allow the atc to start up when credhub is unavailable, create a lazy
credhub client that doesn't talk to credhub until the credhub provider
actually needs to look up credentials.

[Resolves https://github.com/concourse/concourse/issues/2154]